+++ internal/wasi/wasi_test.go
@@ -49,7 +49,7 @@ func TestAPI_ArgsGet(t *testing.T) {
 		3, 0, 0, 0, // little endian-encoded offset of "bc"
 		'?', // stopped after encoding
 	} // tr
-	store, wasiAPI := instantiateWasmStore(t, wasi.FunctionArgsGet, ImportArgsGet, "test", args)
+	store, wasiAPI := instantiateWasmStore(t, FunctionArgsGet, ImportArgsGet, "test", args)
 
 	t.Run("SnapshotPreview1.ArgsGet", func(t *testing.T) {
 		maskMemory(store, maskLength)
@@ -58,10 +58,10 @@ func TestAPI_ArgsGet(t *testing.T) {
 		require.Equal(t, wasi.ErrnoSuccess, errno)
 		require.Equal(t, expectedMemory, store.Memories[0].Buffer[0:maskLength])
 	})
-	t.Run(wasi.FunctionArgsGet, func(t *testing.T) {
+	t.Run(FunctionArgsGet, func(t *testing.T) {
 		maskMemory(store, maskLength)
 
-		ret, _, err := store.CallFunction(context.Background(), "test", wasi.FunctionArgsGet, uint64(argv), uint64(argvBuf))
+		ret, _, err := store.CallFunction(context.Background(), "test", FunctionArgsGet, uint64(argv), uint64(argvBuf))
 		require.NoError(t, err)
 		require.Equal(t, wasi.ErrnoSuccess, wasi.Errno(ret[0])) // cast because results are always uint64
 		require.Equal(t, expectedMemory, store.Memories[0].Buffer[0:maskLength])
@@ -71,7 +71,7 @@ func TestAPI_ArgsGet(t *testing.T) {
 func TestAPI_ArgsGet_Errors(t *testing.T) {
 	args, err := Args("a", "bc")
 	require.NoError(t, err)
-	store, api := instantiateWasmStore(t, wasi.FunctionArgsGet, ImportArgsGet, "test", args)
+	store, api := instantiateWasmStore(t, FunctionArgsGet, ImportArgsGet, "test", args)
 
 	memorySize := uint32(len(store.Memories[0].Buffer))
 	validAddress := uint32(0) // arbitrary valid address as arguments to args_get. We chose 0 here.
@@ -108,7 +108,7 @@ func TestAPI_ArgsGet_Errors(t *testing.T) {
 		tc := tt
 
 		t.Run(tc.name, func(t *testing.T) {
-			ret, _, err := store.CallFunction(context.Background(), "test", wasi.FunctionArgsGet, uint64(tc.argv), uint64(tc.argvBuf))
+			ret, _, err := store.CallFunction(context.Background(), "test", FunctionArgsGet, uint64(tc.argv), uint64(tc.argvBuf))
 			require.NoError(t, err)
 			require.Equal(t, uint64(wasi.ErrnoFault), ret[0]) // ret[0] is returned errno
 		})
@@ -128,7 +128,7 @@ func TestAPI_ArgsSizesGet(t *testing.T) {
 		0x5, 0x0, 0x0, 0x0, // little endian-encoded size of null terminated strings
 		'?', // stopped after encoding
 	} // tr
-	store, wasiAPI := instantiateWasmStore(t, wasi.FunctionArgsSizesGet, ImportArgsSizesGet, "test", args)
+	store, wasiAPI := instantiateWasmStore(t, FunctionArgsSizesGet, ImportArgsSizesGet, "test", args)
 
 	t.Run("SnapshotPreview1.ArgsSizesGet", func(t *testing.T) {
 		maskMemory(store, maskLength)
@@ -138,10 +138,10 @@ func TestAPI_ArgsSizesGet(t *testing.T) {
 		require.Equal(t, wasi.ErrnoSuccess, errno)
 		require.Equal(t, expectedMemory, store.Memories[0].Buffer[0:maskLength])
 	})
-	t.Run(wasi.FunctionArgsSizesGet, func(t *testing.T) {
+	t.Run(FunctionArgsSizesGet, func(t *testing.T) {
 		maskMemory(store, maskLength)
 
-		ret, _, err := store.CallFunction(context.Background(), "test", wasi.FunctionArgsSizesGet, uint64(resultArgc), uint64(resultArgvBufSize))
+		ret, _, err := store.CallFunction(context.Background(), "test", FunctionArgsSizesGet, uint64(resultArgc), uint64(resultArgvBufSize))
 		require.NoError(t, err)
 		require.Equal(t, wasi.ErrnoSuccess, wasi.Errno(ret[0])) // cast because results are always uint64
 		require.Equal(t, expectedMemory, store.Memories[0].Buffer[0:maskLength])
@@ -151,7 +151,7 @@ func TestAPI_ArgsSizesGet(t *testing.T) {
 func TestAPI_ArgsSizesGet_Errors(t *testing.T) {
 	args, err := Args("a", "bc")
 	require.NoError(t, err)
-	store, _ := instantiateWasmStore(t, wasi.FunctionArgsSizesGet, ImportArgsSizesGet, "test", args)
+	store, _ := instantiateWasmStore(t, FunctionArgsSizesGet, ImportArgsSizesGet, "test", args)
 
 	memorySize := uint32(len(store.Memories[0].Buffer))
 	validAddress := uint32(0) // arbitrary valid address as arguments to args_sizes_get. We chose 0 here.
@@ -187,7 +187,7 @@ func TestAPI_ArgsSizesGet_Errors(t *testing.T) {
 		tc := tt
 
 		t.Run(tc.name, func(t *testing.T) {
-			ret, _, err := store.CallFunction(context.Background(), "test", wasi.FunctionArgsSizesGet, uint64(tc.argc), uint64(tc.argvBufSize))
+			ret, _, err := store.CallFunction(context.Background(), "test", FunctionArgsSizesGet, uint64(tc.argc), uint64(tc.argvBufSize))
 			require.NoError(t, err)
 			require.Equal(t, uint64(wasi.ErrnoFault), ret[0]) // ret[0] is returned errno
 		})
@@ -245,7 +245,7 @@ func TestAPI_EnvironGet(t *testing.T) {
 		5, 0, 0, 0, // little endian-encoded offset of "b=cd"
 		'?', // stopped after encoding
 	}
-	store, wasiAPI := instantiateWasmStore(t, wasi.FunctionEnvironGet, ImportEnvironGet, "test", envOpt)
+	store, wasiAPI := instantiateWasmStore(t, FunctionEnvironGet, ImportEnvironGet, "test", envOpt)
 
 	t.Run("API.EnvironGet", func(t *testing.T) {
 		maskMemory(store, maskLength)
@@ -255,10 +255,10 @@ func TestAPI_EnvironGet(t *testing.T) {
 		require.Equal(t, wasi.ErrnoSuccess, errno)
 		require.Equal(t, expectedMemory, store.Memories[0].Buffer[0:maskLength])
 	})
-	t.Run(wasi.FunctionEnvironGet, func(t *testing.T) {
+	t.Run(FunctionEnvironGet, func(t *testing.T) {
 		maskMemory(store, maskLength)
 
-		ret, _, err := store.CallFunction(context.Background(), "test", wasi.FunctionEnvironGet, uint64(resultEnviron), uint64(resultEnvironBuf))
+		ret, _, err := store.CallFunction(context.Background(), "test", FunctionEnvironGet, uint64(resultEnviron), uint64(resultEnvironBuf))
 		require.NoError(t, err)
 		require.Equal(t, wasi.ErrnoSuccess, wasi.Errno(ret[0])) // cast because results are always uint64
 		require.Equal(t, expectedMemory, store.Memories[0].Buffer[0:maskLength])
@@ -268,7 +268,7 @@ func TestAPI_EnvironGet(t *testing.T) {
 func TestAPI_EnvironGet_Errors(t *testing.T) {
 	envOpt, err := Environ("a=bc", "b=cd")
 	require.NoError(t, err)
-	store, api := instantiateWasmStore(t, wasi.FunctionEnvironGet, ImportEnvironGet, "test", envOpt)
+	store, api := instantiateWasmStore(t, FunctionEnvironGet, ImportEnvironGet, "test", envOpt)
 
 	memorySize := uint32(len(store.Memories[0].Buffer))
 	validAddress := uint32(0) // arbitrary valid address as arguments to environ_get. We chose 0 here.
@@ -305,7 +305,7 @@ func TestAPI_EnvironGet_Errors(t *testing.T) {
 		tc := tt
 
 		t.Run(tc.name, func(t *testing.T) {
-			ret, _, err := store.CallFunction(context.Background(), "test", wasi.FunctionEnvironGet, uint64(tc.environ), uint64(tc.environBuf))
+			ret, _, err := store.CallFunction(context.Background(), "test", FunctionEnvironGet, uint64(tc.environ), uint64(tc.environBuf))
 			require.NoError(t, err)
 			require.Equal(t, uint64(wasi.ErrnoFault), ret[0]) // ret[0] is returned errno
 		})
@@ -325,7 +325,7 @@ func TestAPI_EnvironSizesGet(t *testing.T) {
 		0x9, 0x0, 0x0, 0x0, // little endian-encoded size of null terminated strings
 		'?', // stopped after encoding
 	}
-	store, wasiAPI := instantiateWasmStore(t, wasi.FunctionEnvironSizesGet, ImportEnvironSizesGet, "test", envOpt)
+	store, wasiAPI := instantiateWasmStore(t, FunctionEnvironSizesGet, ImportEnvironSizesGet, "test", envOpt)
 
 	t.Run("API.EnvironSizesGet", func(t *testing.T) {
 		maskMemory(store, maskLength)
@@ -335,10 +335,10 @@ func TestAPI_EnvironSizesGet(t *testing.T) {
 		require.Equal(t, wasi.ErrnoSuccess, errno)
 		require.Equal(t, expectedMemory, store.Memories[0].Buffer[0:maskLength])
 	})
-	t.Run(wasi.FunctionEnvironSizesGet, func(t *testing.T) {
+	t.Run(FunctionEnvironSizesGet, func(t *testing.T) {
 		maskMemory(store, maskLength)
 
-		ret, _, err := store.CallFunction(context.Background(), "test", wasi.FunctionEnvironSizesGet, uint64(resultEnvironc), uint64(resultEnvironBufSize))
+		ret, _, err := store.CallFunction(context.Background(), "test", FunctionEnvironSizesGet, uint64(resultEnvironc), uint64(resultEnvironBufSize))
 		require.NoError(t, err)
 		require.Equal(t, wasi.ErrnoSuccess, wasi.Errno(ret[0])) // cast because results are always uint64
 		require.Equal(t, expectedMemory, store.Memories[0].Buffer[0:maskLength])
@@ -349,7 +349,7 @@ func TestAPI_EnvironSizesGet_Errors(t *testing.T) {
 	ctx := context.Background()
 	envOpt, err := Environ("a=b", "b=cd")
 	require.NoError(t, err)
-	store, _ := instantiateWasmStore(t, wasi.FunctionEnvironSizesGet, ImportEnvironSizesGet, "test", envOpt)
+	store, _ := instantiateWasmStore(t, FunctionEnvironSizesGet, ImportEnvironSizesGet, "test", envOpt)
 
 	memorySize := uint32(len(store.Memories[0].Buffer))
 	validAddress := uint32(0) // arbitrary valid address as arguments to environ_sizes_get. We chose 0 here.
@@ -385,7 +385,7 @@ func TestAPI_EnvironSizesGet_Errors(t *testing.T) {
 		tc := tt
 
 		t.Run(tc.name, func(t *testing.T) {
-			ret, _, err := store.CallFunction(ctx, "test", wasi.FunctionEnvironSizesGet, uint64(tc.environc), uint64(tc.environBufSize))
+			ret, _, err := store.CallFunction(ctx, "test", FunctionEnvironSizesGet, uint64(tc.environc), uint64(tc.environBufSize))
 			require.NoError(t, err)
 			require.Equal(t, uint64(wasi.ErrnoFault), ret[0]) // ret[0] is returned errno
 		})
@@ -404,7 +404,7 @@ func TestAPI_ClockTimeGet(t *testing.T) {
 		'?', // stopped after encoding
 	} // tr
 
-	store, api := instantiateWasmStore(t, wasi.FunctionClockTimeGet, ImportClockTimeGet, "test")
+	store, api := instantiateWasmStore(t, FunctionClockTimeGet, ImportClockTimeGet, "test")
 	api.(*wasiAPI).timeNowUnixNano = func() uint64 { return epochNanos }
 
 	t.Run("SnapshotPreview1.ClockTimeGet", func(t *testing.T) {
@@ -415,10 +415,10 @@ func TestAPI_ClockTimeGet(t *testing.T) {
 		require.Equal(t, wasi.ErrnoSuccess, errno)
 		require.Equal(t, expectedMemory, store.Memories[0].Buffer[0:maskLength])
 	})
-	t.Run(wasi.FunctionClockTimeGet, func(t *testing.T) {
+	t.Run(FunctionClockTimeGet, func(t *testing.T) {
 		maskMemory(store, maskLength)
 
-		ret, _, err := store.CallFunction(context.Background(), "test", wasi.FunctionClockTimeGet, 0 /* TODO: id */, 0 /* TODO: precision */, uint64(resultTimestamp))
+		ret, _, err := store.CallFunction(context.Background(), "test", FunctionClockTimeGet, 0 /* TODO: id */, 0 /* TODO: precision */, uint64(resultTimestamp))
 		require.NoError(t, err)
 		require.Equal(t, wasi.ErrnoSuccess, wasi.Errno(ret[0])) // cast because results are always uint64
 		require.Equal(t, expectedMemory, store.Memories[0].Buffer[0:maskLength])
@@ -427,7 +427,7 @@ func TestAPI_ClockTimeGet(t *testing.T) {
 
 func TestAPI_ClockTimeGet_Errors(t *testing.T) {
 	epochNanos := uint64(1640995200000000000) // midnight UTC 2022-01-01
-	store, api := instantiateWasmStore(t, wasi.FunctionClockTimeGet, ImportClockTimeGet, "test")
+	store, api := instantiateWasmStore(t, FunctionClockTimeGet, ImportClockTimeGet, "test")
 	api.(*wasiAPI).timeNowUnixNano = func() uint64 { return epochNanos }
 
 	memorySize := uint32(len(store.Memories[0].Buffer))
@@ -452,7 +452,7 @@ func TestAPI_ClockTimeGet_Errors(t *testing.T) {
 		tc := tt
 
 		t.Run(tc.name, func(t *testing.T) {
-			ret, _, err := store.CallFunction(context.Background(), "test", wasi.FunctionClockTimeGet, 0 /* TODO: id */, 0 /* TODO: precision */, uint64(tc.resultTimestamp))
+			ret, _, err := store.CallFunction(context.Background(), "test", FunctionClockTimeGet, 0 /* TODO: id */, 0 /* TODO: precision */, uint64(tc.resultTimestamp))
 			require.NoError(t, err)
 			require.Equal(t, uint64(wasi.ErrnoFault), ret[0]) // ret[0] is returned errno
 		})
@@ -496,7 +496,7 @@ func TestAPI_ClockTimeGet_Errors(t *testing.T) {
 // TODO: TestAPI_SchedYield TestAPI_SchedYield_Errors
 
 func TestAPI_RandomGet(t *testing.T) {
-	store, api := instantiateWasmStore(t, wasi.FunctionRandomGet, ImportRandomGet, "test")
+	store, api := instantiateWasmStore(t, FunctionRandomGet, ImportRandomGet, "test")
 	maskLength := 7 // number of bytes to write '?' to tell what we've written
 	expectedMemory := []byte{
 		'?',                          // random bytes in `buf` is after this
@@ -526,7 +526,7 @@ func TestAPI_RandomGet(t *testing.T) {
 }
 
 func TestAPI_RandomGet_Errors(t *testing.T) {
-	store, api := instantiateWasmStore(t, wasi.FunctionRandomGet, ImportRandomGet, "test")
+	store, api := instantiateWasmStore(t, FunctionRandomGet, ImportRandomGet, "test")
 
 	memorySize := uint32(len(store.Memories[0].Buffer))
 	validAddress := uint32(0) // arbitrary valid address
@@ -552,7 +552,7 @@ func TestAPI_RandomGet_Errors(t *testing.T) {
 		tc := tt
 
 		t.Run(tc.name, func(t *testing.T) {
-			ret, _, err := store.CallFunction(context.Background(), "test", wasi.FunctionRandomGet, uint64(tc.buf), uint64(tc.bufLen))
+			ret, _, err := store.CallFunction(context.Background(), "test", FunctionRandomGet, uint64(tc.buf), uint64(tc.bufLen))
 			require.NoError(t, err)
 			require.Equal(t, uint64(wasi.ErrnoFault), ret[0]) // ret[0] is returned errno
 		})
